+++ src/App.tsx
@@ -1,16 +1,13 @@
 import React, { useState } from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Dashboard from './pages/Dashboard';
 import ResultsPage from './pages/resultPage'; // Make sure paths are correct
-import MovieTable from "./components/MovieTable";
-import MovieTable2 from "./components/MovieTable2";
-import TVShowTable from "./components/TVShowTable";
-import TVShowTable2 from "./components/TVShowTable2";
 import Header from './components/Header';
 import Footer from './components/Footer';
-// import Dashboard from './pages/dashboard';
-// import { Dashboard } from '@mui/icons-material';
-
+import DashboardPage from './pages/DashboardPage';
+import MovieTopRatedPage from './pages/MovieTopRatedPage';
+import MovieNowPlayingPage from './pages/MovieNowPlayingPage';
+import TVAiringTodayPage from './pages/TVAiringTodayPage';
+import TVTopRatedPage from './pages/TVTopRatedPage';
 const App: React.FC = () => {
   const [selectedCategory, setSelectedCategory] = useState("Top Rated Movies");
   const handleCategorySelect = (category: string) => {
@@ -23,12 +20,12 @@ const App: React.FC = () => {
       
       <Routes>
         {/* Define the routes */}
-        <Route path="/" element={<Dashboard />} />
+        <Route path="/" element={<DashboardPage />} />
         <Route path="/search" element={<ResultsPage />} />
-        <Route path="/movies/top-rated" element={<MovieTable />} />
-        <Route path="/movies/now-playing" element={<MovieTable2 />} />
-        <Route path="/tv/top-rated" element={<TVShowTable />} />
-        <Route path="/tv/airing-today" element={<TVShowTable2 />} />
+        <Route path="/movies/top-rated" element={<MovieTopRatedPage />} />
+        <Route path="/movies/now-playing" element={<MovieNowPlayingPage />} />
+        <Route path="/tv/top-rated" element={<TVTopRatedPage />} />
+        <Route path="/tv/airing-today" element={<TVAiringTodayPage />} />
       </Routes>
 
       <Footer />
