+++ src/components/Header.tsx
@@ -122,6 +122,9 @@ const Header: React.FC<Props> = ({ onCategorySelect }) => {
             <MenuItem onClick={() => handleTVShowsMenuClose("Airing Today TV Shows")}> Airing Today </MenuItem>
           </Menu>
 
+          <Button sx={{ color: "#333"}} color="inherit" > Bookmark </Button>
+          
+
           {/* SEARCHBAR */}
           <Autocomplete
             sx={{ mr: 2, ml: 2, width: "200px" }}
