+++ src/components/TVShowTableAiringToday.tsx
@@ -1,8 +1,16 @@
 import { useEffect, useState } from "react";
-import { createColumnHelper, flexRender, getCoreRowModel, useReactTable } from "@tanstack/react-table";
-import { fetchGenres, fetchTopRatedTV } from "../api/apiConfig";
+import {
+  createColumnHelper,
+  flexRender,
+  getCoreRowModel,
+  useReactTable,
+} from "@tanstack/react-table";
+import {
+  fetchAiringTodayTVShows,
+  fetchGenres,
+} from "../api/apiConfig";
 import axios from "axios";
-import ThumbUpIcon from '@mui/icons-material/ThumbUp';
+import ThumbUpIcon from "@mui/icons-material/ThumbUp";
 
 interface User {
   poster_path: string;
@@ -15,20 +23,42 @@ interface User {
 
 const columnHelper = createColumnHelper<User>();
 
-const columns = (genreMap: { [key: number]: string }, handleLike: (tvId: number) => void) => [
+const columns = (
+  genreMap: { [key: number]: string },
+  handleLike: (tvId: number) => void
+) => [
   columnHelper.accessor("poster_path", {
     header: () => "Poster",
-    cell: (info) => <img src={`https://image.tmdb.org/t/p/w200${info.getValue()}`} alt={info.row.original.title} style={{ height: '100px' }} />
+    cell: (info) => (
+      <img
+        src={`https://image.tmdb.org/t/p/w200${info.getValue()}`}
+        alt={info.row.original.title}
+        style={{ height: "100px" }}
+      />
+    ),
   }),
-  columnHelper.accessor("title", { header: () => "Title", cell: (info) => info.getValue() || "unknown" }),
+  columnHelper.accessor("title", {
+    header: () => "Title",
+    cell: (info) => info.getValue() || "unknown",
+  }),
   columnHelper.accessor("genre_ids", {
     header: () => "Genre",
-    cell: (info) =>{
-        const genres = info.getValue().map((id) => genreMap[id]).join(", ");
-        return genres || "unknown";
-    } 
+    cell: (info) => {
+      const genres = info
+        .getValue()
+        .map((id) => genreMap[id])
+        .filter(Boolean);
+
+      return genres.length > 0 ? genres.join(", ") : "Unknown";
+    },
   }),
-  columnHelper.accessor("vote_average", { header: () => "Rating", cell: (info) => info.getValue() || "unknown" }),
+  columnHelper.accessor("vote_average", {
+    header: () => "Rating",
+    cell: (info) => {
+      const rating = info.getValue();
+      return rating.toFixed(1);
+    },
+  }),
   columnHelper.accessor("like", {
     header: () => "Like",
     cell: (info) => {
@@ -43,7 +73,7 @@ const columns = (genreMap: { [key: number]: string }, handleLike: (tvId: number)
   }),
 ];
 
-const TVShowTable = () => {
+const TVShowTableAiringToday = () => {
   const [tvShows, setTV] = useState<User[]>([]);
   const [genres, setGenres] = useState<{ [key: number]: string }>({});
   const [page, setPage] = useState(1); // Pagination state
@@ -55,10 +85,16 @@ const TVShowTable = () => {
       try {
         setIsLoading(true);
         const fetchedGenres = await fetchGenres();
-        const genreMap = fetchedGenres.reduce((acc: { [key: number]: string }, genre: { id: number, name: string }) => {
-          acc[genre.id] = genre.name;
-          return acc;
-        }, {});
+        const genreMap = fetchedGenres.reduce(
+          (
+            acc: { [key: number]: string },
+            genre: { id: number; name: string }
+          ) => {
+            acc[genre.id] = genre.name;
+            return acc;
+          },
+          {}
+        );
         setGenres(genreMap);
 
         // Fetch the first page of top-rated tv shows
@@ -74,7 +110,7 @@ const TVShowTable = () => {
 
   const fetchTV = async (page: number) => {
     try {
-      const fetchedTV = await fetchTopRatedTV(page); // Pass page and limit 5 items
+      const fetchedTV = await fetchAiringTodayTVShows(page); // Pass page and limit 5 items
       setTV(fetchedTV.results); // Set only current page's results
       setTotalPages(fetchedTV.total_pages); // Set total pages from API response
     } catch (error) {
@@ -117,22 +153,25 @@ const TVShowTable = () => {
   };
 
   return (
-    <div style={{marginTop:"80px", marginBottom:"80px"}} >
-      <h2 >Top Rated TV Shows</h2>
+    <div >
+      <h2>Airing Today TV Shows</h2>
       {isLoading && <p>Loading...</p>}
-      <table  className="users-table"  >
-        <thead >
+      <table className="users-table">
+        <thead>
           {table.getHeaderGroups().map((headerGroup) => (
             <tr key={headerGroup.id}>
               {headerGroup.headers.map((header) => (
                 <th key={header.id} className="users-table-cell">
-                  {flexRender(header.column.columnDef.header, header.getContext())}
+                  {flexRender(
+                    header.column.columnDef.header,
+                    header.getContext()
+                  )}
                 </th>
               ))}
             </tr>
           ))}
         </thead>
-        <tbody >
+        <tbody>
           {table.getRowModel().rows.map((row) => (
             <tr key={row.id}>
               {row.getVisibleCells().map((cell) => (
@@ -144,13 +183,18 @@ const TVShowTable = () => {
           ))}
         </tbody>
       </table>
-      <div className="pagination-controls" >
-        <button onClick={prevPage} disabled={page === 1}> {'<'} </button>
+      <div className="pagination-controls">
+        <button onClick={prevPage} disabled={page === 1}>
+          {" "}
+          {"<"}{" "}
+        </button>
         <span>{`Page ${page} of ${totalPages}`}</span>
-        <button onClick={nextPage} disabled={page === totalPages}>{'>'}</button>
+        <button onClick={nextPage} disabled={page === totalPages}>
+          {">"}
+        </button>
       </div>
     </div>
   );
 };
 
-export default TVShowTable;
+export default TVShowTableAiringToday;
